server/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-celery-beat==2.5.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
prometheus-client==0.19.0
structlog==23.2.0
python-decouple==3.8
whitenoise==6.6.0
django-extensions==3.2.3
